# api/models/embarazada.js
'use strict';
module.exports = (sequelize, DataTypes) => {
  const Embarazada = sequelize.define('Embarazada', {
    imc: DataTypes.DOUBLE,
    escolaridad: DataTypes.STRING,
    semanas_de_gestacion: DataTypes.INTEGER,
    gestas: DataTypes.INTEGER,
    paras: DataTypes.INTEGER,
    abortos: DataTypes.INTEGER,
    cesareas: DataTypes.INTEGER,
    f_ultima_mestruacion: DataTypes.DATEONLY,
    f_probable_de_parto: DataTypes.DATEONLY,
    antecedente_cirugia_uterina: DataTypes.BOOLEAN,
    antecedentes_perinatales: DataTypes.BOOLEAN,
    periodo_intergenesico: DataTypes.BOOLEAN,
    primer_embarazo_o_mas_de_cuatro: DataTypes.BOOLEAN,
    antecedentes_de_dos_o_mas_abortos_espontaneos: DataTypes.BOOLEAN,
    antecedente_cesarea: DataTypes.BOOLEAN,
    preeclampsia: DataTypes.BOOLEAN,
    eclampsia: DataTypes.BOOLEAN,
    obito_fetal: DataTypes.BOOLEAN,
    bajo_peso_al_nacer: DataTypes.BOOLEAN,
    prematurez: DataTypes.BOOLEAN,
    macrosomia: DataTypes.BOOLEAN,
    malformacion_congenita: DataTypes.BOOLEAN,
    sangrado_tercer_trimestre: DataTypes.BOOLEAN,
    polihidramnios: DataTypes.BOOLEAN,
    embarazo_ectopico: DataTypes.BOOLEAN,
    enfermedad_trofoblastica_gestacional: DataTypes.BOOLEAN,
    hipertencion_arterial: DataTypes.BOOLEAN,
    diabetes_mellitus: DataTypes.BOOLEAN,
    cardiopatia_congenita_o_adquirida: DataTypes.BOOLEAN,
    tiroidopatias: DataTypes.BOOLEAN,
    enfermedades_autoinmunes_o_de_la_colagena: DataTypes.BOOLEAN,
    enfermedad_tromboembolica: DataTypes.BOOLEAN,
    accidente_vascular_cerebral: DataTypes.BOOLEAN,
    neumopatias: DataTypes.BOOLEAN,
    tuberculosis: DataTypes.BOOLEAN,
    hepatopatias: DataTypes.BOOLEAN,
    cancer_de_mama: DataTypes.BOOLEAN,
    cancer_cervico_uterino: DataTypes.BOOLEAN,
    cancer_otros_organos: DataTypes.BOOLEAN,
    nefropatias: DataTypes.BOOLEAN,
    epilepsia: DataTypes.BOOLEAN,
    neuropatia_cronica: DataTypes.BOOLEAN,
    hematologicas: DataTypes.BOOLEAN,
    enfermedad_isquemica_miocardica: DataTypes.BOOLEAN,
    discapacidad_fisica_mental: DataTypes.BOOLEAN,
    seropositivo_vih_sida: DataTypes.BOOLEAN,
    alcoholismo: DataTypes.BOOLEAN,
    tabaquismo: DataTypes.BOOLEAN,
    drogadiccion: DataTypes.BOOLEAN,
    adicciones: DataTypes.BOOLEAN,
    f_primera_atencion: DataTypes.DATEONLY,
    f_envio_a_trabajo_social: DataTypes.DATEONLY,
    f_ultima_cita: DataTypes.DATEONLY,
    f_proxima_cita: DataTypes.DATEONLY,
    f_envio_a_cph: DataTypes.DATEONLY,
    f_envio_a_ginecologia: DataTypes.DATEONLY,
    f_envio_de_resolucion_embarazo: DataTypes.DATEONLY,
    f_envio_de_entrega_pulsera_rosa: DataTypes.DATEONLY,
    f_envio_vacula_de_influenza: DataTypes.DATEONLY,
    f_envio_vacula_de_tdpa: DataTypes.DATEONLY,
    f_envio_primera_prueba_vih: DataTypes.DATEONLY,
    f_envio_segunda_prueba_vih: DataTypes.DATEONLY,
    fue_reactiva_alguna_prueba_vih: DataTypes.BOOLEAN,
    ivu: DataTypes.STRING,
    ir: DataTypes.STRING
  }, {});
  Embarazada.associate = function(models) {
    // associations can be defined here
  };
  return Embarazada;
};