# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
#host: localhost
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /embarazadas:
    x-swagger-router-controller: epivector.embarazada.controller

    get:
      operationId: getEmbarazadas
      description: Retorna la lista de entidades
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EmbarazadaList"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"



    post:
      description: Agrega una nueva entidad
      parameters:
        - in: body
          name: entidad
          schema:
            $ref: "#/definitions/Embarazada"
      operationId: createEmbarazada
      responses:
        "201":
          description: success
          schema:
            $ref: "#/definitions/EmbarazadaResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /embarazadas/{id}:
    x-swagger-router-controller: epivector.embarazada.controller

    get:
      description: Retorna una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
      operationId: getEmbarazadaById
      responses:
        "200":
          description: success
          schema:
           $ref: "#/definitions/EmbarazadaResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      description: Actualiza una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
        - in: body
          name: entidad
          schema:
            $ref: "#/definitions/Embarazada"
      operationId: updateEmbarazada
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/EmbarazadaResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: Borra una entidad
      parameters:
        - name: id
          description: Identificador de la entidad
          type: string
          in: path
          required: true
      operationId: deleteEmbarazada
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DeleteResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      code:
        type: number
      message:
        type: string
      description:
        type: string
  #  GeneralResponse:
  #    type: object
  #    properties:
  #      success:
  #        type: number
  #        description: returns 1 if successful
  #      description:
  #        type: string
  #        description: a short comment
  #    required:
  #      - success
  #      - description
...
  Embarazada:
    type: object
    properties:
      imc:
        type: number
        format: double
      escolaridad:
        type: string
      semanas_de_gestacion:
        type: number
      gestas:
        type: number
      paras:
        type: number
      abortos:
        type: number
      cesareas:
        type: number
      f_ultima_mestruacion:
        type: string
        format: date
        example: "2017-01-01"
      f_probable_de_parto:
        type: string
        format: date
        example: "2017-01-01"
      antecedente_cirugia_uterina:
        type: boolean
      antecedentes_perinatales:
        type: boolean
      periodo_intergenesico:
        type: boolean
      primer_embarazo_o_mas_de_cuatro:
        type: boolean
      antecedentes_de_dos_o_mas_abortos_espontaneos:
        type: boolean
      antecedente_cesarea:
        type: boolean
      preeclampsia:
        type: boolean
      eclampsia: 
        type: boolean
      obito_fetal:
        type: boolean
      bajo_peso_al_nacer:
        type: boolean
      prematurez:
        type: boolean
      macrosomia:
        type: boolean
      malformacion_congenita:
        type: boolean
      sangrado_tercer_trimestre:
        type: boolean
      polihidramnios:
        type: boolean
      embarazo_ectopico:
        type: boolean
      enfermedad_trofoblastica_gestacional:
        type: boolean
      hipertencion_arterial:
        type: boolean
      diabetes_mellitus:
        type: boolean
      cardiopatia_congenita_o_adquirida:
        type: boolean
      tiroidopatias:
        type: boolean
      enfermedades_autoinmunes_o_de_la_colagena:
        type: boolean
      enfermedad_tromboembolica:
        type: boolean
      accidente_vascular_cerebral:
        type: boolean
      neumopatias:
        type: boolean
      tuberculosis:
        type: boolean
      hepatopatias:
        type: boolean
      cancer_de_mama:
        type: boolean
      cancer_cervico_uterino:
        type: boolean
      cancer_otros_organos:
        type: boolean
      nefropatias: 
        type: boolean
      epilepsia:
        type: boolean
      neuropatia_cronica:
        type: boolean
      hematologicas:
        type: boolean
      enfermedad_isquemica_miocardica:
        type: boolean
      discapacidad_fisica_mental:
        type: boolean
      seropositivo_vih_sida:
        type: boolean                                   
      alcoholismo:
        type: boolean
      tabaquismo:
        type: boolean
      drogadiccion:
        type: boolean
      adicciones:
        type: boolean
      f_primera_atencion:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_a_trabajo_social:
        type: string
        format: date
        example: "2017-01-01"
      f_ultima_cita:
        type: string
        format: date
        example: "2017-01-01"
      f_proxima_cita:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_a_cph:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_a_ginecologia:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_de_resolucion_embarazo:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_de_entrega_pulsera_rosa:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_vacula_de_influenza:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_vacula_de_tdpa:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_primera_prueba_vih:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_segunda_prueba_vih:
        type: string
        format: date
        example: "2017-01-01"
      fue_reactiva_alguna_prueba_vih:
        type: boolean
      ivu:
        type: string
      ir: 
        type: string
  EmbarazadaResponse:
    type: object
    properties:
      id:
        type: number
      imc:
        type: number
        format: double
      escolaridad:
        type: string
      semanas_de_gestacion:
        type: number
      gestas:
        type: number
      paras:
        type: number
      abortos:
        type: number
      cesareas:
        type: number
      f_ultima_mestruacion:
        type: string
        format: date
        example: "2017-01-01"
      f_probable_de_parto:
        type: string
        format: date
        example: "2017-01-01"
      antecedente_cirugia_uterina:
        type: boolean
      antecedentes_perinatales:
        type: boolean
      periodo_intergenesico:
        type: boolean
      primer_embarazo_o_mas_de_cuatro:
        type: boolean
      antecedentes_de_dos_o_mas_abortos_espontaneos:
        type: boolean
      antecedente_cesarea:
        type: boolean
      preeclampsia:
        type: boolean
      eclampsia: 
        type: boolean
      obito_fetal:
        type: boolean
      bajo_peso_al_nacer:
        type: boolean
      prematurez:
        type: boolean
      macrosomia:
        type: boolean
      malformacion_congenita:
        type: boolean
      sangrado_tercer_trimestre:
        type: boolean
      polihidramnios:
        type: boolean
      embarazo_ectopico:
        type: boolean
      enfermedad_trofoblastica_gestacional:
        type: boolean
      hipertencion_arterial:
        type: boolean
      diabetes_mellitus:
        type: boolean
      cardiopatia_congenita_o_adquirida:
        type: boolean
      tiroidopatias:
        type: boolean
      enfermedades_autoinmunes_o_de_la_colagena:
        type: boolean
      enfermedad_tromboembolica:
        type: boolean
      accidente_vascular_cerebral:
        type: boolean
      neumopatias:
        type: boolean
      tuberculosis:
        type: boolean
      hepatopatias:
        type: boolean
      cancer_de_mama:
        type: boolean
      cancer_cervico_uterino:
        type: boolean
      cancer_otros_organos:
        type: boolean
      nefropatias: 
        type: boolean
      epilepsia:
        type: boolean
      neuropatia_cronica:
        type: boolean
      hematologicas:
        type: boolean
      enfermedad_isquemica_miocardica:
        type: boolean
      discapacidad_fisica_mental:
        type: boolean
      seropositivo_vih_sida:
        type: boolean                                   
      alcoholismo:
        type: boolean
      tabaquismo:
        type: boolean
      drogadiccion:
        type: boolean
      adicciones:
        type: boolean
      f_primera_atencion:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_a_trabajo_social:
        type: string
        format: date
        example: "2017-01-01"
      f_ultima_cita:
        type: string
        format: date
        example: "2017-01-01"
      f_proxima_cita:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_a_cph:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_a_ginecologia:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_de_resolucion_embarazo:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_de_entrega_pulsera_rosa:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_vacula_de_influenza:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_vacula_de_tdpa:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_primera_prueba_vih:
        type: string
        format: date
        example: "2017-01-01"
      f_envio_segunda_prueba_vih:
        type: string
        format: date
        example: "2017-01-01"
      fue_reactiva_alguna_prueba_vih:
        type: boolean
      ivu:
        type: string
      ir: 
        type: string
  EmbarazadaList:
    properties:
      embarazadas:
        type: array
        items:
          $ref: "#/definitions/EmbarazadaResponse"
    required:
      - embarazadas
  DeleteResponse:
    type: object
    properties:
      success:
        type: number
        description: Retorna 1 en caso de exito
      description:
        type: string
        description: Comentario
    required:
      - success
      - description
...